requirements.txt
Django==4.0.2
django-filter==21.1
djangorestframework==3.13.1
Markdown==3.3.6
requests==2.27.1
jsonpath==0.82


